requirements.txt
absl-py==0.5.0
ale-python-interface==0.6
astor==0.7.1
atari-py==0.1.6
backcall==0.1.0
bleach==1.5.0
cached-property==1.3.0
certifi==2018.10.15
chainer==1.8.1
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
filelock==3.0.12
future==0.16.0
gast==0.2.0
grpcio==1.15.0
gym==0.15.4
gym-pull==0.1.7
gym-tetris==3.0.2
h5py==2.8.0
html5lib==0.9999999
idna==2.7
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
nes-py==8.1.1
networkx==2.2
nose==1.3.7
numpy==1.15.2
opencv-python==4.1.1.26
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==3.0.2
protobuf==3.6.1
ptyprocess==0.6.0
pydot==1.4.1
pyglet==1.3.2
Pygments==2.5.2
pyparsing==2.2.2
python-dateutil==2.7.3
pytouch==0.1.0
PyWavelets==1.0.1
PyYAML==3.13
requests==2.19.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
tensorboard==1.6.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
toolz==0.9.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.23
wcwidth==0.1.8
Werkzeug==0.14.1
